feat: launch of vjoule dumper core, and download of the traces

+++ web-apps/socialNet/deploy/src/deployer/cluster.cc
@@ -14,8 +14,9 @@ namespace deployer {
                         auto user = (*machines) [mc]["user"].getStr ();
                         auto iface = (*machines) [mc].getOr ("iface", "eth0");
                         auto workDir = (*machines) [mc].getOr ("work-dir", "~/");
+                        auto hasPDU = (*machines) [mc].getOr ("has-pdu", false);
 
-                        this-> _machines.emplace (mc, std::make_shared <Machine> (hostname, user, workDir, iface));
+                        this-> _machines.emplace (mc, std::make_shared <Machine> (hostname, user, workDir, iface, hasPDU));
                     }
                 } catch (const std::runtime_error & err) {
                     LOG_ERROR ("Malformed configuration file : ", err.what ());
@@ -58,6 +59,46 @@ namespace deployer {
     /*!
      * ====================================================================================================
      * ====================================================================================================
+     * ====================================          PREPARE          =====================================
+     * ====================================================================================================
+     * ====================================================================================================
+     */
+
+    void Cluster::prepareVJoule () {
+        for (auto & it : this-> _machines) {
+            if (it.second-> hasFlag ("social") || it.second-> hasFlag ("cache")) {
+                LOG_INFO ("Starting vjoule on ", it.first);
+                it.second-> run ("systemctl stop vjoule_service")-> wait ();
+                it.second-> run ("systemctl start vjoule_service")-> wait ();
+            }
+        }
+    }
+
+    void Cluster::downloadVJouleTraces (const std::string & resultDir) {
+        for (auto & it : this-> _machines) {
+            if (it.second-> hasFlag ("social") || it.second-> hasFlag ("cache")) {
+                auto dir = rd_utils::utils::join_path (resultDir, it.first);
+                rd_utils::utils::create_directory (dir, true);
+
+                auto cgroupPath = rd_utils::utils::join_path (dir, "cgroups.csv");
+                auto energyPath = rd_utils::utils::join_path (dir, "energy.csv");
+
+                try {
+                    it.second-> get ("/etc/vjoule/results/cgroups.csv", cgroupPath);
+                } catch (...) {}
+
+                try {
+                    it.second-> get ("/etc/vjoule/results/energy.csv", energyPath);
+                } catch (...) {}
+
+                it.second-> run ("systemctl stop vjoule_service")-> wait ();
+            }
+        }
+    }
+
+    /*!
+     * ====================================================================================================
+     * ====================================================================================================
      * ===================================          DISPOSING          ====================================
      * ====================================================================================================
      * ====================================================================================================
@@ -66,10 +107,15 @@ namespace deployer {
     void Cluster::clean () {
         for (auto & mch : this-> _machines) {
             LOG_INFO ("Removing temp directories on ", mch.first);
-
             mch.second-> run ("rm -rf /tmp/" + mch.first + "*")-> wait ();
-            auto runDir = rd_utils::utils::join_path (mch.second-> getHomeDir (), "run");
-            mch.second-> run ("rm -rf " + runDir)-> wait ();
+
+            if (mch.second-> hasFlag ("social") || mch.second-> hasFlag ("cache")) {
+                auto runDir = rd_utils::utils::join_path (mch.second-> getHomeDir (), "run");
+                mch.second-> run ("rm -rf " + runDir)-> wait ();
+
+                LOG_INFO ("Stopping vjoule on ", mch.first);
+                mch.second-> run ("systemctl stop vjoule_service")-> wait ();
+            }
         }
     }
